Major Refactor

+++ src/config.rs
@@ -7,7 +7,7 @@ pub struct Config {
     //fs::File,
     pub init_with: String,
     //fs::File,
-    pub radius_scale: u16,
+    pub radius_scale: f64,
     pub resolution: u16,
     pub max_depth: u16,
 }
@@ -18,7 +18,7 @@ impl Config {
             load_from: "".to_string(), //fs::File(),
             save_at: "".to_string(),   //fs::File(),
             init_with: "".to_string(), //fs::File(),
-            radius_scale: 1,
+            radius_scale: 1.0,
             resolution: 256,
             max_depth: 4,
         })
@@ -30,7 +30,7 @@ impl Config {
             load_from: "".to_string(), //fs::File(),
             save_at: "".to_string(),   //fs::File(),
             init_with: "".to_string(), //fs::File(),
-            radius_scale: 1,
+            radius_scale: 1.0,
             resolution: 256,
             max_depth: 4,
         })
